Make score an int
See the note to Eval.cpp

+++ Eval.h
@@ -1,5 +1,5 @@
 #ifndef _Eval_H_Included_
 #define _Eval_H_Included_
 
-float Evaluate_Position();
+int Evaluate_Position();
 #endif
